Introduce new loggers for testing and reproduction number

+++ cpp/models/abm/infection.cpp
@@ -47,9 +47,8 @@ Infection::Infection(Person::RandomNumberGenerator& rng, VirusVariant virus, Age
         vl_params.viral_load_incline.get_distribution_instance()(rng, vl_params.viral_load_incline.params);
     m_viral_load.decline =
         vl_params.viral_load_decline.get_distribution_instance()(rng, vl_params.viral_load_decline.params);
-    m_viral_load.end_date =
-        m_viral_load.start_date +
-        days(int(m_viral_load.peak / m_viral_load.incline - m_viral_load.peak / m_viral_load.decline));
+    m_viral_load.end_date = m_viral_load.start_date +
+                            days(m_viral_load.peak / m_viral_load.incline - m_viral_load.peak / m_viral_load.decline);
 
     auto inf_params = params.get<InfectivityDistributions>()[{virus, age}];
     m_log_norm_alpha =
